refactor: more handle errors func to the errors.py
